refactor: manage 패키지와 menu 패키지를 만들어 비슷한 클래스들끼리 묶음
