+++ src/main/java/subway/Application.java
@@ -1,13 +1,15 @@
 package subway;
 
 import java.util.Scanner;
-import subway.controller.Subway;
+import subway.controller.SubwayController;
+import subway.view.InputView;
 
 public class Application {
 
     public static void main(String[] args) {
         final Scanner scanner = new Scanner(System.in);
-        Subway subway = new Subway(scanner);
+        final InputView inputView = new InputView(scanner);
+        SubwayController subway = new SubwayController(inputView);
         subway.start();
     }
 }
