+++ src/main/java/subway/exception/SubwayCustomException.java
@@ -1,14 +1,12 @@
 package subway.exception;
-
-import subway.view.OutputView;
-
 /**
  * 커스텀 예외 처리 클래스
  */
 public class SubwayCustomException extends IllegalArgumentException {
 
+    public static final String ERROR_PREFIX = "[ERROR] ";
+
     public SubwayCustomException(String message) {
-        super(OutputView.showErrorMessage(message));
-        System.out.println();
+        super(ERROR_PREFIX + message);
     }
 }
