+++ src/main/java/subway/menu/LineMenu.java
@@ -1,21 +1,22 @@
-package subway.view;
+package subway.menu;
 
-import static subway.view.MainMenu.WRONG_MENU_INPUT_EXCEPTION_MESSAGE;
+import static subway.view.TextFixtures.WRONG_MENU_INPUT_EXCEPTION_MESSAGE;
 
 import java.util.Arrays;
-import subway.controller.Management;
+import subway.controller.manager.LineManager;
 import subway.exception.SubwayCustomException;
+import subway.view.OutputView;
 
-public enum StationMenu {
-    ADD_STATION("1", Management::addStation),
-    DELETE_STATION("2", Management::deleteStation),
-    SHOW_STATION_LIST("3", OutputView::showStationList),
+public enum LineMenu {
+    ADD_STATION("1", LineManager::addLine),
+    DELETE_STATION("2", LineManager::deleteLine),
+    SHOW_STATION_LIST("3", OutputView::showLineList),
     GO_BACK("B", null);
 
     private final String input;
     private final Runnable handler;
 
-    StationMenu(String input, Runnable handler) {
+    LineMenu(String input, Runnable handler) {
         this.input = input;
         this.handler = handler;
     }
@@ -28,6 +29,7 @@ public enum StationMenu {
                 .orElseThrow(() -> new SubwayCustomException(WRONG_MENU_INPUT_EXCEPTION_MESSAGE))
                 .handler.run();
         } catch (NullPointerException exception) {
+
         }
     }
 }
