+++ src/main/java/subway/menu/MainMenu.java
@@ -1,18 +1,23 @@
-package subway.view;
+package subway.menu;
+
+import static subway.view.TextFixtures.WRONG_MENU_INPUT_EXCEPTION_MESSAGE;
 
 import java.util.Arrays;
-import subway.controller.Management;
+import subway.controller.manager.LineManager;
+import subway.controller.manager.MainManager;
+import subway.controller.manager.SectionManager;
+import subway.controller.manager.StationManager;
 import subway.exception.SubwayCustomException;
+import subway.view.OutputView;
 
 public enum MainMenu {
-    MANAGE_STATION("1", Management::manageStation),
-    MANAGE_LINE("2", Management::manageLine),
-    MANAGE_SECTION("3", Management::mangeSection),
+    MANAGE_STATION("1", StationManager::manageStation),
+    MANAGE_LINE("2", LineManager::manageLine),
+    MANAGE_SECTION("3", SectionManager::mangeSection),
     MANAGE_SUBWAY("4", OutputView::showSubwayLineMap),
-    MANAGE_QUIT("Q", MainMenu::end);
+    MANAGE_QUIT("Q", MainManager::end);
 
 
-    public static final String WRONG_MENU_INPUT_EXCEPTION_MESSAGE = "선택할 수 없는 기능입니다.";
     private final String input;
     private final Runnable handler;
 
@@ -27,10 +32,6 @@ public enum MainMenu {
             .findFirst()
             .orElseThrow(() -> new SubwayCustomException(WRONG_MENU_INPUT_EXCEPTION_MESSAGE))
             .handler.run();
-    }
-
-    private static void end() {
-        System.exit(0);
     }
 }
